+++ pubspec.yaml
@@ -3,7 +3,7 @@ description: A new Flutter project.
 
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+publish_to: "none" # Remove this line if you wish to publish to pub.dev
 
 # The following defines the version and build number for your application.
 # A version number is three numbers separated by dots, like 1.2.43
@@ -20,7 +20,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.19.0-146.2.beta <3.0.0'
+  sdk: ">=2.19.0-146.2.beta <3.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -29,31 +29,25 @@ environment:
 # the latest version available on pub.dev. To see which dependencies have newer
 # versions available, run `flutter pub outdated`.
 dependencies:
-  cupertino_icons: ^1.0.5
-  firebase_storage: ^11.1.0
-  firebase_core: ^2.9.0
-  firebase_auth: ^4.2.2
-  google_sign_in: ^6.1.0
   cloud_firestore: ^4.5.0
-  get: ^4.6.5
-  image_picker: ^0.8.7+1
-  font_awesome_flutter: ^10.4.0
-  shared_preferences: ^2.1.0
-  table_calendar: ^3.0.9
-  intl: ^0.18.0
+  cupertino_icons: ^1.0.5
+  firebase_auth: ^4.2.2
+  firebase_core: ^2.9.0
+  firebase_storage: ^11.1.0
   flutter:
     sdk: flutter
-  flutter_svg: ^2.0.5
-  flutter_native_splash: ^2.2.19
+  flutter_native_splash: ^2.2.18
   flutter_pdfview: ^1.3.0
+  flutter_svg: ^1.1.6
+  font_awesome_flutter: ^10.4.0
+  get: ^4.6.5
+  google_sign_in: ^6.1.0
+  image_picker: ^0.8.7+2
+  intl: ^0.18.0
+  shared_preferences: ^2.1.0
+  table_calendar: ^3.0.9
 
 dev_dependencies:
-
-  # The "flutter_lints" package below contains a set of recommended lints to
-  # encourage good coding practices. The lint set provided by the package is
-  # activated in the `analysis_options.yaml` file located at the root of your
-  # package. See that file for information about deactivating specific lint
-  # rules and activating additional ones.
   flutter_lints: ^2.0.1
   flutter_test:
     sdk: flutter
@@ -62,7 +56,6 @@ dev_dependencies:
 # following page: https://dart.dev/tools/pub/pubspec
 # The following section is specific to Flutter packages.
 flutter:
-
   # The following line ensures that the Material Icons font is
   # included with your application, so that you can use the icons in
   # the material Icons class.
@@ -72,7 +65,7 @@ flutter:
     - assets/logo.svg
     - assets/explanation_1.svg
     - assets/explanation_2.svg
-    - assets/explanation_3.svg  
+    - assets/explanation_3.svg
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware
   # For details regarding adding assets from package dependencies, see
